pyproject.toml

[tool.poetry]
name = "abitur-ai"
version = "0.1.0"
description = ""
authors = ["Sergey Dmitriev <51058739+0niel@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
html2text = "^2024.2.26"
langchain = "^0.2.3"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
progress = "^1.6"
pydantic = "^2.7.3"
langchain-google-genai = "^1.0.6"
langchain-google-vertexai = {git = "https://github.com/langchain-ai/langchain-google.git", subdirectory = "libs/vertexai"}
python-dotenv = "^1.0.1"
langchain-core = "^0.2.10"
langchain-openai = "^0.1.10"
python-telegram-bot = "^21.3"
pydantic-settings = "^2.3.4"


[tool.poetry.group.dev.dependencies]
setuptools = "^70.0.0"
mypy = "^1.10.0"
black = "^24.4.2"
isort = "^5.13.2"
autoflake = "^2.3.1"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 120
